pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mosquito_object_detection"
version = "0.0.1"
authors = [
  { name="Brandon Knape", email="d.brandonk@gmail.com" },
]
description = "mosquito_object_detection"
readme = "README.md"
requires-python = ">=3.7"

dependencies = [
    'aicrowd-cli',
    'numpy',
    'pandas',
    'pyyaml',
    'ultralytics',
]
